add struct: DirPath
