+++ sources/apps/bisque/src/command/download_file.rs
@@ -1,8 +1,8 @@
+use bisque_core::models::{DirPath, FilePath};
 use bisque_core::AccessTokenLoader;
 use bisque_core::{command::download_file, BisqueClient};
 use bisque_google_drive::schemas::FileName;
 use clap::Parser;
-use std::path::PathBuf;
 
 #[derive(Parser, Debug)]
 pub struct Args {
@@ -28,10 +28,10 @@ pub fn run(args: Args) -> crate::Result<()> {
 
     let client = BisqueClient::new(access_token)?;
     client.download_file(download_file::Params {
-        key_file_path: args.key_file,
+        key_file_path: FilePath::verify(args.key_file)?,
         src_name: FileName::new(args.file_name.clone())?,
         src_folder_id: args.folder_id,
-        dst_dir_path: PathBuf::from(args.download_dir),
+        dst_dir_path: DirPath::verify(args.download_dir)?,
     })?;
     Ok(())
 }
